chore: split out cli and api targets
