+++ src/choices/groups/index.ts
@@ -17,7 +17,8 @@
 import { Node, Element } from "hast"
 import { visit } from "unist-util-visit"
 
-import { ChoiceState, MadWizardOptions } from "../.."
+import { ChoiceState } from "../../choices"
+import { MadWizardOptions } from "../../fe/MadWizardOptions"
 import { isTabGroup } from "../../parser/markdown/rehype-tabbed"
 export { getTabTitle, isTabWithProperties, setTabGroup, setTabTitle } from "../../parser/markdown/rehype-tabbed"
